refactor: Improve quality of MindeeClient usage
