+++ src/main/java/com/mindee/MindeeClient.java
@@ -1,7 +1,6 @@
 package com.mindee;
 
 import com.mindee.parsing.MindeeApi;
-import com.mindee.parsing.MindeeSettings;
 import com.mindee.parsing.PageOptions;
 import com.mindee.parsing.common.Document;
 import com.mindee.parsing.common.Inference;
@@ -19,13 +18,16 @@ public class MindeeClient {
   private final MindeeApi mindeeApi;
   private final PdfOperation pdfOperation;
 
-  public MindeeClient() {
-    this(new MindeeSettings());
-  }
-
-  public MindeeClient(MindeeSettings mindeeSettings) {
-    this.mindeeApi = new MindeeApi(mindeeSettings);
+  public MindeeClient(
+    MindeeApi mindeeApi) {
+    this.mindeeApi = mindeeApi;
     this.pdfOperation = new PdfBoxApi();
+  }
+  public MindeeClient(
+    MindeeApi mindeeApi,
+    PdfOperation pdfOperation) {
+    this.mindeeApi = mindeeApi;
+    this.pdfOperation = pdfOperation;
   }
 
   public DocumentToParse loadDocument(InputStream fileStream,
@@ -81,7 +83,7 @@ public class MindeeClient {
     return this.mindeeApi.predict(
       type,
       ParseParameter.builder()
-        .file(documentToParse.getFile())
+        .file(getSplitFile(documentToParse, pageOptions))
         .fileName(documentToParse.getFilename())
         .build());
   }
@@ -92,6 +94,16 @@ public class MindeeClient {
     boolean includeWords,
     PageOptions pageOptions) throws IOException {
 
+    return this.mindeeApi.predict(
+      type,
+      ParseParameter.builder()
+        .file(getSplitFile(documentToParse, pageOptions))
+        .fileName(documentToParse.getFilename())
+        .includeWords(includeWords)
+        .build());
+  }
+
+  private byte[] getSplitFile(DocumentToParse documentToParse, PageOptions pageOptions) throws IOException {
     byte[] splitFile;
 
     if (FileUtils.getFileExtension(documentToParse.getFilename())
@@ -102,12 +114,6 @@ public class MindeeClient {
       splitFile = documentToParse.getFile();
     }
 
-    return this.mindeeApi.predict(
-      type,
-      ParseParameter.builder()
-        .file(splitFile)
-        .fileName(documentToParse.getFilename())
-        .includeWords(includeWords)
-        .build());
+    return splitFile;
   }
 }
