+++ src/test/java/com/mindee/parsing/MindeeClientTest.java
@@ -4,42 +4,26 @@ import com.mindee.DocumentToParse;
 import com.mindee.MindeeClient;
 import com.mindee.parsing.common.Document;
 import com.mindee.parsing.invoice.InvoiceV4Inference;
-import okhttp3.mockwebserver.MockResponse;
-import okhttp3.mockwebserver.MockWebServer;
-import org.junit.jupiter.api.AfterEach;
+import com.mindee.pdf.PdfOperation;
+import com.mindee.pdf.SplitPdf;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import javax.print.Doc;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.List;
 
 @ExtendWith(MockitoExtension.class)
 class MindeeClientTest {
 
   MindeeClient client;
-  MockWebServer mockWebServer = new MockWebServer();
-
-  @BeforeEach
-  void setUp() throws IOException {
-    mockWebServer.start();
-    client = new MindeeClient(
-      new MindeeSettings(
-        "abc",
-        String.format("http://localhost:%s", mockWebServer.getPort())));
-  }
-
-  @AfterEach
-  public void tearDown() throws Exception {
-    mockWebServer.shutdown();
-  }
+  MindeeApi mindeeApi;
+  PdfOperation pdfOperation;
 
   @Test
   void givenAClientWithMissingConfigured_whenParsed_thenShouldThrowException() {
@@ -49,28 +33,105 @@ class MindeeClientTest {
   }
 
   @Test
-  void givenAClientWithInvoiceConfigured_whenParsed_thenShouldCallTheHttpClientCorrectly()
+  void givenAClientForInvoice_withFile_parse_thenShouldCallMindeeApi()
     throws IOException {
 
-    Path path = Paths.get("src/test/resources/data/invoice/response_v4/complete.json");
-    mockWebServer.enqueue(new MockResponse()
-      .setResponseCode(200)
-      .setBody(new String(Files.readAllBytes(path))));
-
+    mindeeApi = Mockito.mock(MindeeApi.class);
+    client = new MindeeClient(mindeeApi);
     File file = new File("src/test/resources/data/invoice/invoice.pdf");
+
+    Mockito.when(
+        mindeeApi.predict(
+          Mockito.any(),
+          Mockito.any()))
+      .thenReturn(new Document<>());
+
     Document<InvoiceV4Inference> document = client.parse(
       InvoiceV4Inference.class,
       new DocumentToParse(file));
 
     Assertions.assertNotNull(document);
+  }
 
-    String[] actualLines = document.toString().split(System.lineSeparator());
-    String actualSummary = String.join(String.format("%n"), actualLines);
-    List<String> expectedLines = Files
-      .readAllLines(Paths.get("src/test/resources/data/invoice/response_v4/summary.txt"));
-    String expectedSummary = String.join(String.format("%n"), expectedLines);
+  @Test
+  void givenAClientForInvoice_withInputStream_parse_thenShouldCallMindeeApi()
+    throws IOException {
+
+    mindeeApi = Mockito.mock(MindeeApi.class);
+    client = new MindeeClient(mindeeApi);
+    File file = new File("src/test/resources/data/invoice/invoice.pdf");
+
+    Mockito.when(
+        mindeeApi.predict(
+          Mockito.any(),
+          Mockito.any()))
+      .thenReturn(new Document<>());
+
+    Document<InvoiceV4Inference> document = client.parse(
+      InvoiceV4Inference.class,
+      new DocumentToParse(
+        Files.newInputStream(file.toPath()),
+        ""));
 
     Assertions.assertNotNull(document);
-    Assertions.assertEquals(expectedSummary, actualSummary);
+  }
+
+  @Test
+  void givenAClientForInvoice_withByteArray_parse_thenShouldCallMindeeApi()
+    throws IOException {
+
+    mindeeApi = Mockito.mock(MindeeApi.class);
+    client = new MindeeClient(mindeeApi);
+    File file = new File("src/test/resources/data/invoice/invoice.pdf");
+
+    Mockito.when(
+        mindeeApi.predict(
+          Mockito.any(),
+          Mockito.any()))
+      .thenReturn(new Document<>());
+
+    Document<InvoiceV4Inference> document = client.parse(
+      InvoiceV4Inference.class,
+      new DocumentToParse(
+        Files.readAllBytes(file.toPath()),
+        ""));
+
+    Assertions.assertNotNull(document);
+    Mockito.verify(mindeeApi, Mockito.times(1))
+      .predict(Mockito.any(),Mockito.any());
+  }
+
+  @Test
+  void givenAClientForInvoiceAndPageOptions_parse_thenShouldOperateCutOnPagesAndCallTheHttpClientCorrectly()
+    throws IOException {
+
+    mindeeApi = Mockito.mock(MindeeApi.class);
+    pdfOperation = Mockito.mock(PdfOperation.class);
+    client = new MindeeClient(mindeeApi, pdfOperation);
+    File file = new File("src/test/resources/data/invoice/invoice.pdf");
+    List<Integer> pageNumberToKeep = new ArrayList<>();
+    pageNumberToKeep.add(1);
+
+    Mockito.when(
+        mindeeApi.predict(
+          Mockito.any(),
+          Mockito.any()))
+      .thenReturn(new Document<>());
+    Mockito.when(
+        pdfOperation.split(
+          Mockito.any()))
+      .thenReturn(new SplitPdf(new byte[0], 0));
+
+    Document<InvoiceV4Inference> document = client.parse(
+      InvoiceV4Inference.class,
+      new DocumentToParse(file),
+      new PageOptions(
+        pageNumberToKeep, PageOptionsOperation.KEEP_ONLY_LISTED_PAGES, 0));
+
+    Assertions.assertNotNull(document);
+    Mockito.verify(mindeeApi, Mockito.times(1))
+      .predict(Mockito.any(),Mockito.any());
+    Mockito.verify(pdfOperation, Mockito.times(1))
+      .split(Mockito.any());
   }
 }
